requirements.txt
astroid==2.1.0
cairocffi==0.9.0
CairoSVG==2.2.1
certifi==2018.10.15
cffi==1.11.5
cssselect2==0.2.1
defusedxml==0.5.0
Django==2.1.2
django-cors-headers==2.4.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
html5lib==1.0.1
isort==4.3.4
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
Pillow==5.4.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycparser==2.19
PyJWT==1.6.4
pylint==2.2.2
Pyphen==0.9.5
pytz==2018.7
redis==3.0.1
reportlab==3.5.12
six==1.12.0
tinycss2==0.6.1
typed-ast==1.1.1
WeasyPrint==44
webencodings==0.5.1
wrapt==1.10.11
